Add CustomRevisionFilter module to handle shared logic between ArticleScopedProgram and VisitingScholarship.
